Extract Sidebar component
