+++ html-report/src/App.vue
@@ -1,38 +1,22 @@
 <script setup lang="ts">
 import { ref } from 'vue';
-import { ResizableConfig, vResizable } from 'vue-resizables'
 import ExecutionDetails from './components/details/ExecutionDetails.vue';
-import ExecutionTree from './components/sidebar/ExecutionTree.vue';
 import TestNodeDetails from './components/details/TestNodeDetails.vue';
 import Selection from './Selection';
 import StatusBar from './components/header/StatusBar.vue';
 import TestExecution from './TestExecution';
-import ToolBar from './components/sidebar/ToolBar.vue';
+import Sidebar from "./components/sidebar/Sidebar.vue";
 
 const props = defineProps<{ executions: TestExecution[] }>()
 const selection = ref<Selection | undefined>();
 selection.value = TestExecution.initialSelection(props.executions)
-const resizeConfig: ResizableConfig = {
-  edge: {
-    right: true,
-  },
-  size: {
-    min: {
-      width: 320, // min-w-80
-    }
-  }
-}
 </script>
 
 <template>
   <main class="flex flex-col h-screen bg-white dark:bg-black text-black dark:text-white">
     <StatusBar :executions="executions" />
     <div class="flex flex-1 overflow-hidden">
-      <div v-resizable="resizeConfig"
-        class="resize-x bg-neutral-100 dark:bg-neutral-800 w-1/2 lg:w-1/3 2xl:w-1/4 min-w-80 overflow-auto text-sm border-r border-neutral-200 dark:border-neutral-700 h-full">
-        <ToolBar :executions="executions" />
-        <ExecutionTree :executions="executions" v-model:selection="selection" class="ml-1.5 mt-2.5" />
-      </div>
+      <Sidebar :executions="executions" v-model:selection="selection" class="overflow-auto w-1/2 lg:w-1/3 2xl:w-1/4 border-r border-neutral-200 dark:border-neutral-700 h-full" />
       <div class="flex-1 overflow-auto">
         <ExecutionDetails v-if="(selection?.item instanceof TestExecution)" :execution="selection?.item" />
         <TestNodeDetails :execution="selection?.execution" :node="(selection?.item as TestNodeData)"
@@ -41,5 +25,3 @@ const resizeConfig: ResizableConfig = {
     </div>
   </main>
 </template>
-
-<style scoped></style>
