+++ html-report/src/components/sidebar/Sidebar.vue
@@ -1,0 +1,29 @@
+<script setup lang="ts">
+import {ResizableConfig, vResizable} from 'vue-resizables'
+import ExecutionTree from "./ExecutionTree.vue";
+import ToolBar from "./ToolBar.vue";
+import TestExecution from "../../TestExecution.ts";
+import Selection from "../../Selection.ts";
+
+const selection = defineModel<Selection | undefined>('selection')
+defineProps<{ executions: TestExecution[] }>()
+
+const resizeConfig: ResizableConfig = {
+  edge: {
+    right: true,
+  },
+  size: {
+    min: {
+      width: 320, // min-w-80
+    }
+  }
+}
+</script>
+
+<template>
+  <div v-resizable="resizeConfig"
+       class="resize-x bg-neutral-100 dark:bg-neutral-800 min-w-80 text-sm">
+    <ToolBar :executions="executions" />
+    <ExecutionTree :executions="executions" v-model:selection="selection" class="ml-1.5 mt-2.5" />
+  </div>
+</template>
